+++ tests/Sources/CommandLineArgumentsSource.test.ts
@@ -1,5 +1,5 @@
 import "jest";
-import { CommandLineSource } from "../../src/Sources/CommandLineSource";
+import { CommandLineArgumentsSource } from "../../src/Sources/CommandLineArgumentsSource";
 import { ConfigurationSourceError } from "../../src/Sources/ConfigurationSourceError";
 import { ConfigurationSourceInterface } from "../../src/Sources/ConfigurationSourceInterface";
 
@@ -9,7 +9,7 @@ test("should check if an argument name is not empty", () => {
         "/app.js",
         "run",
     ];
-    expect(() => new CommandLineSource("", args))
+    expect(() => new CommandLineArgumentsSource("", args))
         .toThrow(new ConfigurationSourceError("The command line argument name can not be empty."));
 });
 
@@ -40,7 +40,7 @@ test("should process override arguments", () => {
         "--override",
         "zeta=eta",
     ];
-    const cliSource: ConfigurationSourceInterface = new CommandLineSource("override", args);
+    const cliSource: ConfigurationSourceInterface = new CommandLineArgumentsSource("override", args);
     expect(cliSource.resolve())
         .toEqual({
             top1: "level1",
@@ -72,7 +72,7 @@ test("should throw an error if argument is invalid", () => {
         ["/bin/node", "/app.js", "--override=", "--whatever"],
         ["/bin/node", "/app.js", "--override", "--whatever"],
     ]) {
-        const cliSource: ConfigurationSourceInterface = new CommandLineSource("override", args);
+        const cliSource: ConfigurationSourceInterface = new CommandLineArgumentsSource("override", args);
         expect(() => cliSource.resolve())
             .toThrow(new ConfigurationSourceError(`One or more "--override" command line arguments does not define configuration item name.`));
     }
@@ -81,7 +81,7 @@ test("should throw an error if argument is invalid", () => {
         ["/bin/node", "/app.js", "--override=whatever"],
         ["/bin/node", "/app.js", "--override", "whatever"],
     ]) {
-        const cliSource: ConfigurationSourceInterface = new CommandLineSource("override", args);
+        const cliSource: ConfigurationSourceInterface = new CommandLineArgumentsSource("override", args);
         expect(() => cliSource.resolve())
             .toThrow(new ConfigurationSourceError(`One or more "--override" command line arguments does not define configuration item value.`));
     }
@@ -95,7 +95,7 @@ test("should process one argument (--change-this)", () => {
         "--override=epsilon=zeta",
         "run",
     ];
-    const cliSource: ConfigurationSourceInterface = new CommandLineSource("change-this", args);
+    const cliSource: ConfigurationSourceInterface = new CommandLineArgumentsSource("change-this", args);
     expect(cliSource.resolve())
         .toEqual({
             alpha: "beta",
@@ -108,7 +108,7 @@ test("should handle no arguments", () => {
         "/app.js",
         "run",
     ];
-    const cliSource: ConfigurationSourceInterface = new CommandLineSource("override", args);
+    const cliSource: ConfigurationSourceInterface = new CommandLineArgumentsSource("override", args);
     expect(cliSource.resolve())
         .toEqual({});
 });